Refactoring / revert reaction module to package in main module
