--- reactions/src/main/AndroidManifest.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="im.vector.reactions">
-
-    <application>
-        <meta-data
-            android:name="fontProviderRequests"
-            android:value="Noto Color Emoji Compat" />
-
-        <activity
-            android:name=".EmojiReactionPickerActivity"
-            android:label="@string/title_activity_emoji_reaction_picker"
-            android:theme="@style/Theme.AppCompat.Light.NoActionBar"
-            android:exported="true"
-            >
-        </activity>
-    </application>
-
-</manifest>
