--- reactions/src/main/res/drawable/circle.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="oval">
-    <solid android:color="@color/pale_grey" />
-</shape>
