--- reactions/src/main/res/layout/activity_emoji_reaction_picker.xml
@@ -1,68 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.coordinatorlayout.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".EmojiReactionPickerActivity">
-
-
-    <fragment
-        android:id="@+id/fragment"
-        android:name="im.vector.reactions.EmojiChooserFragment"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        app:layout_behavior="@string/appbar_scrolling_view_behavior"
-        tools:layout="@layout/emoji_chooser_fragment" />
-    <!--<androidx.core.widget.NestedScrollView-->
-    <!--android:id="@+id/nestedScrollView"-->
-    <!--android:layout_width="match_parent"-->
-    <!--android:layout_height="match_parent"-->
-    <!--android:fillViewport="true"-->
-    <!--android:fitsSystemWindows="true"-->
-    <!--app:layout_behavior="@string/appbar_scrolling_view_behavior">-->
-
-    <!--<fragment-->
-    <!--android:id="@+id/fragment"-->
-    <!--android:name="im.vector.reactions.EmojiChooserFragment"-->
-    <!--android:layout_width="match_parent"-->
-    <!--android:layout_height="match_parent"-->
-    <!--tools:layout="@layout/emoji_chooser_fragment" />-->
-
-    <!--</androidx.core.widget.NestedScrollView>-->
-
-    <com.google.android.material.appbar.AppBarLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content">
-
-        <!--<com.google.android.material.appbar.CollapsingToolbarLayout-->
-        <!--android:id="@+id/collapsingToolbarLayout"-->
-        <!--android:layout_width="match_parent"-->
-        <!--android:layout_height="wrap_content"-->
-        <!--app:contentScrim="?attr/colorPrimary"-->
-        <!--app:expandedTitleGravity="top"-->
-        <!--app:layout_scrollFlags="scroll|exitUntilCollapsed|snap">-->
-
-        <androidx.appcompat.widget.Toolbar
-            android:id="@+id/toolbar"
-            android:layout_width="match_parent"
-            android:layout_height="?attr/actionBarSize"
-            android:background="?attr/colorPrimary"
-            android:elevation="4dp"
-            android:minHeight="0dp"
-            android:theme="@style/ThemeOverlay.AppCompat.ActionBar"
-            app:layout_scrollFlags="scroll|exitUntilCollapsed|snap|enterAlways"
-            app:popupTheme="@style/ThemeOverlay.AppCompat.Light" />
-
-        <com.google.android.material.tabs.TabLayout
-            android:id="@+id/tabs"
-            android:layout_width="match_parent"
-            android:layout_height="40dp"
-            android:background="?attr/colorPrimary"
-            android:elevation="4dp" />
-
-        <!--</com.google.android.material.appbar.CollapsingToolbarLayout>-->
-    </com.google.android.material.appbar.AppBarLayout>
-
-
-</androidx.coordinatorlayout.widget.CoordinatorLayout>
