--- reactions/src/main/res/layout/grid_item_emoji.xml
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="40dp"
-    android:layout_height="40dp"
-    tools:showIn="@layout/activity_emoji_reaction_picker">
-
-    <View
-        android:id="@+id/grid_item_place_holder"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_gravity="center"
-        android:layout_margin="4dp"
-        android:visibility="gone"
-        tools:visibility="visible"
-        android:background="@drawable/circle" />
-
-    <im.vector.reactions.EmojiDrawView
-        android:id="@+id/grid_item_emoji_text"
-        android:layout_width="40dp"
-        android:layout_height="40dp"
-        android:layout_gravity="center" />
-    <!--<TextView-->
-    <!--android:layout_gravity="center"-->
-    <!--android:id="@+id/grid_item_emoji_text"-->
-    <!--android:layout_width="wrap_content"-->
-    <!--android:layout_height="wrap_content"-->
-    <!--tools:text="😀"-->
-    <!--android:textSize="24sp"-->
-    <!--/>-->
-</FrameLayout>
