--- reactions/src/main/res/layout/grid_section_header.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content">
-
-    <TextView
-        android:id="@+id/section_header_textview"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_margin="8dp"
-        android:paddingStart="8dp"
-        android:paddingLeft="8dp"
-        android:paddingRight="8dp"
-        android:textStyle="bold"
-        tools:text="Smiley &amp; Peolple" />
-
-</LinearLayout>
