--- reactions/src/main/res/values/colors.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <color name="pale_grey">#f2f5f8</color>
-    <color name="light_blue_grey">#4ac1c9d6</color>
-</resources>
