--- reactions/src/main/res/values/strings.xml
@@ -1,4 +1,0 @@
-<resources>
-    <string name="app_name">reactions</string>
-    <string name="title_activity_emoji_reaction_picker">Reactions</string>
-</resources>
