+++ vector/src/main/res/layout/emoji_chooser_fragment.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.recyclerview.widget.RecyclerView xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/emoji_recycler_view"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:scrollbars="vertical"
+    tools:context="im.vector.riotredesign.features.reactions.EmojiChooserFragment"
+    tools:itemCount="100"
+    tools:layoutManager="androidx.recyclerview.widget.GridLayoutManager"
+    tools:listitem="@layout/grid_item_emoji"
+    tools:spanCount="10">
+
+</androidx.recyclerview.widget.RecyclerView>
